feat: move GlobalState to separate module
