+++ crates/window/src/lib.rs
@@ -1,33 +1,22 @@
 mod cursors;
 pub mod event;
+mod global;
 mod hook;
 mod message_loop;
 mod types;
 mod window;
 
 use core::{
-    ptr,
+    error::Error,
+    fmt::Display,
     sync::atomic::{AtomicBool, Ordering},
 };
 
 use anyhow::Context;
 use asdf_overlay_common::cursor::Cursor;
 use once_cell::sync::OnceCell;
-use parking_lot::RwLock;
-use windows::Win32::{
-    Foundation::RECT,
-    UI::{
-        Input::Ime::ImmCreateContext,
-        WindowsAndMessaging::{GetSystemMetrics, SM_CXVIRTUALSCREEN, SM_CYVIRTUALSCREEN},
-    },
-};
 
-use crate::{
-    event::{BackendEvent, WindowEvent},
-    message_loop::MessageLoopState,
-    types::IntDashMap,
-    window::WindowProcState,
-};
+use crate::{event::BackendEvent, global::GlobalState};
 
 static GLOBAL: OnceCell<GlobalState> = OnceCell::new();
 
@@ -44,18 +33,7 @@ impl Backends {
             hook::install()?;
             message_loop::hook::install()?;
 
-            let _blocking_ime_cx = unsafe { ImmCreateContext() }.0 as usize;
-            Ok(GlobalState {
-                hinstance,
-
-                event_tx,
-
-                message_loops: IntDashMap::default(),
-                windows: IntDashMap::default(),
-
-                blocking_state: RwLock::new(None),
-                blocking_cursor: RwLock::new(Some(Cursor::Default)),
-            })
+            Ok(GlobalState::new(hinstance, event_tx))
         };
 
         static INITIALIZED: AtomicBool = AtomicBool::new(false);
@@ -67,6 +45,18 @@ impl Backends {
             .get_or_try_init(init_inner)
             .context("initialization failed")?;
         Ok(Self { event_rx })
+    }
+
+    pub fn recv(&self) -> Option<BackendEvent> {
+        self.event_rx.recv().ok()
+    }
+
+    pub async fn recv_async(&self) -> Option<BackendEvent> {
+        self.event_rx.recv_async().await.ok()
+    }
+
+    pub fn try_recv(&self) -> Result<BackendEvent, TryRecvError> {
+        Ok(self.event_rx.try_recv()?)
     }
 
     #[inline]
@@ -106,146 +96,28 @@ impl Drop for Backends {
     }
 }
 
-pub(crate) struct GlobalState {
-    hinstance: usize,
-
-    event_tx: flume::Sender<BackendEvent>,
-
-    message_loops: IntDashMap<u32, MessageLoopState>,
-    windows: IntDashMap<u32, WindowProcState>,
-
-    blocking_cursor: RwLock<Option<Cursor>>,
-    blocking_state: RwLock<Option<InputBlockingState>>,
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum TryRecvError {
+    Empty,
+    Disconnected,
 }
 
-impl GlobalState {
-    pub(crate) fn input_blocked(&self) -> bool {
-        self.blocking_state.read().is_some()
-    }
-
-    pub(crate) fn blocking_cursor(&self) -> Option<Cursor> {
-        *self.blocking_cursor.read()
-    }
-
-    /// Block input for the window.
-    pub fn block_input(&self) {
-        let clip_cursor = {
-            let mut rect = RECT::default();
-            let global_hook = hook::HOOK.wait();
-
-            unsafe {
-                _ = global_hook.get_clip_cursor.original_fn()(&mut rect);
-                let screen = RECT {
-                    left: 0,
-                    top: 0,
-                    right: GetSystemMetrics(SM_CXVIRTUALSCREEN),
-                    bottom: GetSystemMetrics(SM_CYVIRTUALSCREEN),
-                };
-                _ = global_hook.clip_cursor.original_fn()(ptr::null());
-
-                if rect != screen { Some(rect) } else { None }
-            }
-        };
-
-        for message_loop in self.message_loops.iter() {
-            message_loop.block_input();
+impl From<flume::TryRecvError> for TryRecvError {
+    fn from(err: flume::TryRecvError) -> Self {
+        match err {
+            flume::TryRecvError::Empty => TryRecvError::Empty,
+            flume::TryRecvError::Disconnected => TryRecvError::Disconnected,
         }
-
-        for window in self.windows.iter() {
-            window.block_input();
-        }
-
-        *self.blocking_state.write() = Some(InputBlockingState { clip_cursor });
-    }
-
-    /// Unblock input for the window.
-    pub fn unblock_input(&self) {
-        for message_loop in self.message_loops.iter() {
-            message_loop.unblock_input();
-        }
-
-        for window in self.windows.iter() {
-            window.unblock_input();
-        }
-
-        *self.blocking_state.write() = None;
-        self.emit(BackendEvent::InputBlockingEnded);
-    }
-
-    /// Get or initialize the message loop state for the given thread ID.
-    ///
-    /// NOTE: The thread id is windows system thread id, not rust thread id.
-    fn message_loop_state<R>(&self, thread_id: u32, f: impl FnOnce(&MessageLoopState) -> R) -> R {
-        if let Some(state) = self.message_loops.get(&thread_id) {
-            return f(state.value());
-        }
-
-        let state = self
-            .message_loops
-            .entry(thread_id)
-            .or_insert_with(|| {
-                let state = MessageLoopState::new(thread_id);
-                if self.input_blocked() {
-                    state.block_input();
-                }
-
-                state
-            })
-            .downgrade();
-        f(state.value())
-    }
-
-    fn cleanup_message_loop(&self, thread_id: u32) {
-        self.message_loops.remove(&thread_id);
-    }
-
-    fn window_state<R>(&self, window_id: u32, f: impl FnOnce(&WindowProcState) -> R) -> R {
-        if let Some(state) = self.windows.get(&window_id) {
-            return f(state.value());
-        }
-
-        let state = self
-            .windows
-            .entry(window_id)
-            .or_try_insert_with(|| -> anyhow::Result<WindowProcState> {
-                let state = WindowProcState::init(window_id)?;
-
-                let (width, height) = state.size();
-                self.emit(BackendEvent::Window {
-                    id: window_id,
-                    event: WindowEvent::Added { width, height },
-                });
-
-                if self.input_blocked() {
-                    state.block_input();
-                }
-
-                Ok(state)
-            })
-            .expect("failed to initialize window state")
-            .downgrade();
-        f(state.value())
-    }
-
-    fn cleanup_window(&self, window_id: u32) {
-        if self.windows.remove(&window_id).is_none() {
-            return;
-        }
-
-        self.emit(BackendEvent::Window {
-            id: window_id,
-            event: WindowEvent::Destroyed,
-        });
-    }
-
-    /// Emit [`BackendEvent`] to event sink. If one exists.
-    #[inline]
-    pub(crate) fn emit(&self, event: BackendEvent) {
-        _ = self.event_tx.send(event);
     }
 }
 
-struct InputBlockingState {
-    // Old cursor clipping rectangle, if any.
-    clip_cursor: Option<RECT>,
+impl Display for TryRecvError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            TryRecvError::Empty => flume::TryRecvError::Empty.fmt(f),
+            TryRecvError::Disconnected => flume::TryRecvError::Disconnected.fmt(f),
+        }
+    }
 }
+
+impl Error for TryRecvError {}
